Resolve #333 -- Buffs have a slot value now
The slot value is hardcoded to be 1 right now. We'll need to change that
in the future when adding a buff manager that assigns buffs a free slot

Refs #333

+++ GameServerLib/Logic/Packets/Packets.cs
@@ -1921,7 +1921,7 @@ namespace LeagueSandbox.GameServer.Logic.Packets
 
     public class AddBuff : BasePacket
     {
-        public AddBuff(Unit u, Unit source, int stacks, float time, BuffType buffType, string name, int slot)
+        public AddBuff(Unit u, Unit source, int stacks, float time, BuffType buffType, string name, byte slot)
                : base(PacketCmd.PKT_S2C_AddBuff, u.NetId)
         {
             var _rafManager = Program.ResolveDependency<RAFManager>();
